Fixed data handlers not creating files when saving

+++ src/main/java/host/serenity/serenity/files/accounts/AccountDataHandler.java
@@ -13,6 +13,10 @@ public class AccountDataHandler implements ClientDataHandler {
 
     @Override
     public void save() throws IOException {
+        if (!file.exists() && !file.createNewFile()) {
+            return;
+        }
+
         BufferedWriter writer = new BufferedWriter(new FileWriter(file));
 
         for (Account account : SerenityPluginAltManager.getAccountManager().getAccounts()) {
